rename directory + use suffix instead of prefix for private variables
